workable but incredibly complicated changing rotations overtime for everything - started with 1 cube

+++ src/mygame/CustomIndividualCube.java
@@ -5,9 +5,9 @@ import com.jme3.scene.Spatial;
 public class CustomIndividualCube {
 
 	Spatial spatialObject;
-	private int rotX;
-	private int rotY;
-	private int rotZ;
+	private float rotX;
+	private float rotY;
+	private float rotZ;
 	public CustomIndividualCube(Spatial spatialObject){
 		this.spatialObject = spatialObject;
 	}
@@ -17,29 +17,31 @@ public class CustomIndividualCube {
 		
 	}
 
-	public int getRotX() {
+	public float getRotX() {
 		return rotX;
 	}
 
-	public void setRotX(int rotX) {
+	public void setRotX(float rotX) {
 		this.rotX = rotX;
 	}
 
-	public int getRotY() {
+	public float getRotY() {
 		return rotY;
 	}
 
-	public void setRotY(int rotY) {
+	public void setRotY(float rotY) {
 		this.rotY = rotY;
 	}
 
-	public int getRotZ() {
+	public float getRotZ() {
 		return rotZ;
 	}
 
-	public void setRotZ(int rotZ) {
+	public void setRotZ(float rotZ) {
 		this.rotZ = rotZ;
 	}
+
+	
 	
 	
 }
